move flash notice to controller concern
The FindCreatePerson module has been converted
to a concern called SessionPersonCreator since it
is a mixin fundamentally (not a service object).
Further it can now be mixed into the TokenController
and SessionController, thereby enabling its internals
to have access to controller helpers.
